+++ src/ebm/losses/__init__.py
@@ -1,5 +1,6 @@
 from ebm.losses.base import LossOutput
 from ebm.losses.cd import ContrastiveDivergence
+from ebm.losses.discrete import PseudoLikelihood, RatioMatching
 from ebm.losses.drl import DiffusionRecoveryLikelihood, drl_sample
 from ebm.losses.energy_discrepancy import EnergyDiscrepancy
 from ebm.losses.nce import NoiseContrastiveEstimation
@@ -17,6 +18,8 @@ __all__ = [
     "DiffusionRecoveryLikelihood",
     "drl_sample",
     "EnergyDiscrepancy",
+    "PseudoLikelihood",
+    "RatioMatching",
     "DenoisingScoreMatching",
     "ExactScoreMatching",
     "MultiSigmaDenoisingScoreMatching",